requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Werkzeug==3.0.1
python-dotenv==1.0.0
Pillow==10.1.0
pytesseract==0.3.10
openai==1.3.0
stripe==7.4.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
requests==2.31.0
